introduce and use apk_blob_fmt/apk_fmt/apk_fmts instead of snprintf
While there are no obvious paths to trigger any misbehaviour this
removes usage of snprintf and potential bad patterns. Error handling
in few instances is added as defensive means.

fixes #11030

+++ src/fs_fsys.c
@@ -252,7 +252,8 @@ static int fsys_file_control(struct apk_fsdir *d, apk_blob_t filename, int ctrl)
 		break;
 	case APK_FS_CTRL_APKNEW:
 		// rename tmpname -> realname.apk-new
-		snprintf(apknewname, sizeof apknewname, "%s%s", fn, ".apk-new");
+		rc = apk_fmt(apknewname, sizeof apknewname, "%s%s", fn, ".apk-new");
+		if (rc < 0) break;
 		if (renameat(atfd, format_tmpname(&ac->dctx, d->pkgctx, dirname, apk_pathbuilder_get(&d->pb), tmpname),
 			     atfd, apknewname) < 0)
 			rc = -errno;
@@ -269,7 +270,8 @@ static int fsys_file_control(struct apk_fsdir *d, apk_blob_t filename, int ctrl)
 		break;
 	case APK_FS_CTRL_DELETE_APKNEW:
 		// remove apknew (which may or may not exist)
-		snprintf(apknewname, sizeof apknewname, "%s%s", fn, ".apk-new");
+		rc = apk_fmt(apknewname, sizeof apknewname, "%s%s", fn, ".apk-new");
+		if (rc < 0) break;
 		unlinkat(atfd, apknewname, 0);
 		break;
 	default:
